Move device related resize into separate files
Try to keep the rest of the ConfigSpec resize from getting
unwieldy.
